WO/ps1/pyproject.toml
[project]
name = "ps1"
version = "0.1.0"
description = "Sniffing Womans PussyAndAss"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-dotenv>=1.0.1",
    "rich>=13.9.4",
    "supabase>=2.11.0",
]
